Move quill modules to modules directory
