+++ modules/emojis/src/emoji-list.js
