+++ modules/emojis/src/emoji-map.js
