+++ modules/emojis/src/format-emoji-blot.js
