+++ modules/emojis/src/module-emoji.js
