+++ modules/emojis/src/module-textarea-emoji.js
