+++ modules/emojis/src/module-toolbar-emoji.js
