+++ modules/emojis/src/quill-emoji.js
