+++ modules/emojis/src/scss/core/_emoji.scss
