+++ modules/emojis/src/scss/core/_styles.scss
