+++ modules/emojis/src/scss/core/_variables.scss
