+++ modules/emojis/src/scss/quill-emoji.scss
