+++ modules/emojis/src/sheet_apple_16.png
