+++ modules/emojis/src/sheet_apple_20.png
