+++ modules/emojis/src/sheet_apple_32.png
