+++ modules/quill-mention/src/constants/index.js
