+++ modules/table/src/assets/quill-better-table.scss
