+++ modules/table/src/modules/table-selection.js
